some additions

See merge request djakonystar911/ElectroLight-Warehouse!24

+++ app/src/main/java/uz/texnopos/elektrolife/data/model/newsale/Product.kt
@@ -16,11 +16,11 @@ data class Product(
     @SerializedName("product_cost_price")
     val productCostPrice: Int,
     @SerializedName("price_wholesale")
-    val priceWholesale: Long,
+    val priceWholesale: Double,
     @SerializedName("price_min")
-    val priceMin: Long,
+    val priceMin: Double,
     @SerializedName("price_max")
-    val priceMax: Long,
+    val priceMax: Double,
     val remained: Int,
     var count: Int,
     var salePrice: Long
